refactor: change structure, remove class componets
