+++ react-project/src/App.tsx
@@ -1,11 +1,16 @@
-import ErrorBoundary from "./components/ErrorBoundary";
 import "./App.css";
 import Layout from "./components/layout/Layout";
+import Search from "./components/search-nav/Search";
+import ErrorBoundary from "./components/error-boundary/ErrorBoundary";
+import Results from "./components/search-results/Results";
 
 function App() {
   return (
     <ErrorBoundary>
-      <Layout>Main content</Layout>
+      <Layout>
+        <Search />
+        <Results />
+      </Layout>
     </ErrorBoundary>
   );
 }
