+++ react-project/src/components/book-card/Card.tsx
