+++ react-project/src/components/button/Button.tsx
@@ -1,7 +1,9 @@
 type ButtonProps = {
   children: React.ReactNode;
-  className: string;
-  handleClick: (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void;
+  className?: string;
+  handleClick?: (
+    event: React.MouseEvent<HTMLButtonElement, MouseEvent>,
+  ) => void;
 };
 
 function Button({ children, className, handleClick }: ButtonProps) {
