+++ react-project/src/components/search-results/Results.tsx
@@ -1,0 +1,9 @@
+function Results() {
+  return (
+    <div className="Results">
+      <h1>Results</h1>
+    </div>
+  );
+}
+
+export default Results;
